requirements.txt
boto3
python-dotenv
scalebox-sdk
langchain
langchain-aws